story updates
