+++ stories/Parallax/3_ParallaxInADiv.stories.js
@@ -3,36 +3,13 @@ import { text, number } from '@storybook/addon-knobs';
 import { storiesOf } from '@storybook/react';
 import { Parallax, ParallaxProvider } from 'react-scroll-parallax';
 import Element from '../Element/Element';
-
+import ScrollContainer from '../ScrollContainer';
 import styles from './Parallax.scss';
 import '../styles.scss';
 
-class ScrollContainer extends React.Component {
-    constructor() {
-        super();
-        this.scrollContainer = React.createRef();
-        this.state = {
-            mounted: false,
-        };
-    }
-
-    componentDidMount = () => {
-        this.setState({ mounted: true });
-    };
-
-    render() {
-        const ref = this.scrollContainer.current;
-        return (
-            <div className="scroll-container" ref={this.scrollContainer}>
-                {this.state.mounted ? this.props.children(ref) : null}
-            </div>
-        );
-    }
-}
-
 storiesOf('<Parallax> in a <div>', module)
     .add('vertical with x offsets', () => {
-        const amount = number('Number of Elements', 3);
+        const amount = number('Number of Elements', 2);
         const offA = number('x[0] * i', -50);
         const offB = number('x[1] * i', 50);
         const unit = text('Offset Unit', '%');
@@ -66,7 +43,7 @@ storiesOf('<Parallax> in a <div>', module)
         );
     })
     .add('vertical with y offsets', () => {
-        const amount = number('Number of Elements', 3);
+        const amount = number('Number of Elements', 2);
         const offA = number('y[0] * i', -50);
         const offB = number('y[1] * i', 50);
         const unit = text('Offset Unit', '%');
@@ -100,7 +77,7 @@ storiesOf('<Parallax> in a <div>', module)
         );
     })
     .add('horizontal with x offsets', () => {
-        const amount = number('Number of Elements', 3);
+        const amount = number('Number of Elements', 2);
         const offA = number('x[0] * i', -50);
         const offB = number('x[1] * i', 50);
         const unit = text('Offset Unit', '%');
@@ -137,7 +114,7 @@ storiesOf('<Parallax> in a <div>', module)
         );
     })
     .add('horizontal with y offsets', () => {
-        const amount = number('Number of Elements', 3);
+        const amount = number('Number of Elements', 2);
         const offA = number('y[0] * i', -50);
         const offB = number('y[1] * i', 50);
         const unit = text('Offset Unit', '%');
